+++ packages/surveycto/test/index.js
@@ -1,10 +1,9 @@
 import { expect } from 'chai';
 
-import nock from 'nock';
 import ClientFixtures, { fixtures } from './ClientFixtures';
 
 import Adaptor from '../src';
-const { execute, event, dataElement, get } = Adaptor;
+const { execute } = Adaptor;
 
 describe('execute', () => {
   it.skip('executes each operation in sequence', done => {
@@ -39,38 +38,6 @@ describe('execute', () => {
         references: [],
         data: null,
       });
-    });
-  });
-});
-
-describe('get', () => {
-  before(() => {
-    nock('https://play.http.org')
-      .get('/demo/api/events')
-      .reply(200, { foo: 'bar' });
-  });
-
-  it.skip('calls the callback', () => {
-    let state = {
-      configuration: {
-        username: 'hello',
-        password: 'there',
-        baseUrl: 'https://play.http.org/demo',
-      },
-    };
-
-    return execute(
-      get('api/events', {
-        callback: (response, state) => {
-          return { ...state, references: [response] };
-        },
-        username: null,
-      })
-    )(state).then(state => {
-      let lastReference = state.references[0];
-
-      // Check that the eventData made it's way to the request as a string.
-      expect(lastReference).to.eql({ foo: 'bar' });
     });
   });
 });