+++ packages/surveycto/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @openfn/language-surveycto
 
+## 2.0.0
+
+### Major Changes
+
+- 59ae7b50: - Refactor `fetchSubmissions()` function
+  - Update `instanceName` to `servername` in `configuration-schema.json`
+  - Add `apiVersion` in `configuration-schema.json`. Defaults to `v1`
+  - Add `request()` function for surveyCTO API requests
+
 ## 1.2.1
 
 ### Patch Changes